+++ NotificareMonetizeKit/Sources/Internals/Database/MonetizeDatabase.swift
@@ -8,11 +8,16 @@ import NotificareKit
 
 internal class MonetizeDatabase: NotificareAbstractDatabase {
     init() {
-        super.init(name: "NotificareMonetizeDatabase", rebuildOnVersionChange: true)
-        context.mergePolicy = NSOverwriteMergePolicy
+        super.init(
+            name: "NotificareMonetizeDatabase",
+            rebuildOnVersionChange: true,
+            mergePolicy: .overwrite
+        )
     }
 
     internal func add(_ purchase: NotificarePurchase) throws -> PurchaseEntity {
+        ensureLoadedStores()
+
         let entity = try PurchaseEntity(context, purchase: purchase)
         saveChanges()
 
@@ -20,6 +25,8 @@ internal class MonetizeDatabase: NotificareAbstractDatabase {
     }
 
     internal func find() throws -> [PurchaseEntity] {
+        ensureLoadedStores()
+
         let request = NSFetchRequest<PurchaseEntity>(entityName: "PurchaseEntity")
         let result = try context.fetch(request)
 
@@ -27,6 +34,8 @@ internal class MonetizeDatabase: NotificareAbstractDatabase {
     }
 
     internal func clear() throws {
+        ensureLoadedStores()
+
         let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: "PurchaseEntity")
         let deleteRequest = NSBatchDeleteRequest(fetchRequest: fetchRequest)