+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 3.6.1
+
+- Fix race condition when setting the database merge policy eagerly loads the data stores
+- Prevent fatal error when failing to open the databases
+
 ## 3.6.0
 
 - Allow checking which regions are being monitored